[UPDATE] Log창에 진행 시간 보여주는 기능 추가

+++ Assets/01.Script/Metronome.cs
@@ -37,7 +37,7 @@ public class Metronome : MonoBehaviour
     }
 
     private void Update(){
-        (audioSource.timeSamples / audioSource.clip.frequency).Log();
+        ((float)audioSource.timeSamples / (float)audioSource.clip.frequency).Log();
         if(audioSource.timeSamples >= nextSample){
             StartCoroutine(TicSFX());
         }
